update rendering & creation APIs
